+++ easynav_simple_common/src/easynav_simple_common/SimpleMap.cpp
@@ -26,9 +26,7 @@
 #include <sstream>
 
 #include "nav_msgs/msg/occupancy_grid.hpp"
-#include "geometry_msgs/msg/pose.hpp"
 
-#include "easynav_common/types/MapTypeBase.hpp"
 #include "easynav_simple_common/SimpleMap.hpp"
 
 namespace easynav
@@ -93,8 +91,8 @@ SimpleMap::check_bounds_metric(double mx, double my) const
     return false;
   }
 
-  int x = static_cast<int>(relative_x / resolution_);
-  int y = static_cast<int>(relative_y / resolution_);
+  size_t x = static_cast<size_t>(relative_x / resolution_);
+  size_t y = static_cast<size_t>(relative_y / resolution_);
 
   return x < width_ && y < height_;
 }
@@ -139,7 +137,7 @@ SimpleMap::to_occupancy_grid(nav_msgs::msg::OccupancyGrid & grid_msg) const
     grid_msg.data.resize(width_ * height_);
   }
 
-  for (int idx = 0; idx < data_.size(); ++idx) {
+  for (size_t idx = 0; idx < data_.size(); ++idx) {
     grid_msg.data[idx] = data_[idx] ? 100 : 0;
   }
 }
@@ -155,9 +153,9 @@ SimpleMap::from_occupancy_grid(const nav_msgs::msg::OccupancyGrid & grid_msg)
     grid_msg.info.origin.position.y,
     false);  // valor inicial: libre (false)
 
-  for (int y = 0; y < height_; ++y) {
-    for (int x = 0; x < width_; ++x) {
-      int idx = y * width_ + x;
+  for (size_t y = 0; y < height_; ++y) {
+    for (size_t x = 0; x < width_; ++x) {
+      size_t idx = y * width_ + x;
       int8_t val = grid_msg.data[idx];
       data_[idx] = (val >= 50);
     }
@@ -176,7 +174,7 @@ SimpleMap::save_to_file(const std::string & path) const
       << resolution_ << " "
       << origin_x_ << " " << origin_y_ << "\n";
 
-  for (int i = 0; i < data_.size(); ++i) {
+  for (size_t i = 0; i < data_.size(); ++i) {
     out << (data_[i] ? "1" : "0");
     if (i + 1 < data_.size()) {
       out << " ";
@@ -200,7 +198,7 @@ SimpleMap::load_from_file(const std::string & path)
 
   if (!std::getline(in, line)) {return false;}
   std::istringstream meta_stream(line);
-  int w, h;
+  size_t w, h;
   double res, ox, oy;
   if (!(meta_stream >> w >> h >> res >> ox >> oy)) {
     return false;
@@ -236,7 +234,7 @@ SimpleMap::index(int x, int y) const
 }
 
 void
-SimpleMap::check_bounds(int x, int y) const
+SimpleMap::check_bounds(size_t x, size_t y) const
 {
   if (x >= width_ || y >= height_) {
     throw std::out_of_range("SimpleMap: index out of bounds");
@@ -259,8 +257,8 @@ SimpleMap::print(bool view_data) const
   if (!view_data) {return;}
 
   std::cerr << "SimpleMap Data:\n";
-  for (int y = 0; y < height_; ++y) {
-    for (int x = 0; x < width_; ++x) {
+  for (size_t y = 0; y < height_; ++y) {
+    for (size_t x = 0; x < width_; ++x) {
       double mx = origin_x_ + (x + 0.5) * resolution_;
       double my = origin_y_ + (y + 0.5) * resolution_;
       std::cerr << "[" << x << ", " << y << "][" << mx << ", " << my << "] "
